requirements.txt
flask==3.0.3
flask-cors==4.0.1
pymongo==4.7.3
pandas==2.2.2
scikit-learn==1.4.2
scipy==1.13.1
numpy==1.26.4
python-dotenv==1.0.1
